+++ src/GenBuild/AssemblyInfo.fs
@@ -1,0 +1,14 @@
+﻿namespace System
+open System.Reflection
+
+[<assembly: AssemblyTitleAttribute("GenBuild")>]
+[<assembly: AssemblyProductAttribute("GenBuild")>]
+[<assembly: AssemblyCompanyAttribute("halcwb")>]
+[<assembly: AssemblyDescriptionAttribute("A library with utility functions")>]
+[<assembly: AssemblyVersionAttribute("0.0.1")>]
+[<assembly: AssemblyFileVersionAttribute("0.0.1")>]
+do ()
+
+module internal AssemblyVersionInformation =
+    let [<Literal>] Version = "0.0.1"
+    let [<Literal>] InformationalVersion = "0.0.1"