+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 #### 0.0.1-alpha - May 22 2016
+* **Note: This is a dummy library to test the use of a general build script!**
 * Initial setup